% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mesos_qmd_files.R
\name{create_mesos_qmd_files}
\alias{create_mesos_qmd_files}
\title{Create QMD-files for the mesos groups in a mesos variable}
\usage{
create_mesos_qmd_files(dir_path, mesos_var, mesos_groups)
}
\arguments{
\item{dir_path}{String, path to folder where the reference qmd-files are stored
(and where in which the new mesos folders will be created)}

\item{mesos_var}{String, inserted into qmd-files as \code{params$mesos_var}}

\item{mesos_groups}{Character vector of mesos groups.}
}
\value{
Data frame with with names.
}
\description{
Create QMD-files for the mesos groups in a mesos variable
}
\examples{
writeLines(c(""), con = fs::path(tempdir(), "index.qmd"))
create_mesos_qmd_files(dir_path=tempdir(),
mesos_var = "f_uni", mesos_groups = paste0("Uni of ", LETTERS[1:5]))
}
