% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_reports_from_sidebars.R
\name{remove_reports_from_sidebars}
\alias{remove_reports_from_sidebars}
\title{Remove Reports from Sidebars}
\usage{
remove_reports_from_sidebars(
  path = "_site",
  filename_as_regex = c("0_report\\\\.pdf", "0_report\\\\.docx")
)
}
\arguments{
\item{path}{Folder path as string.}

\item{filename_as_regex}{Character vector of regex filenames to look for (i.e.,
the dot separating file extension must be escaped as \\.).}
}
\value{
List of files processed.
}
\description{
Finds HTML-files recursively in a path, and removes all entries in the sidebar
to certain filenames, such as "0_report.pdf"
}
\examples{
remove_reports_from_sidebars(path = tempdir())
}
